나의 즐겨 찾기 부분 스코어 자료형 double로 변경

+++ searchStudy/src/mypage/model/dao/StoreDao.java
@@ -91,7 +91,7 @@ public class StoreDao{
 					stores.setStoreName(rset.getString("store_name"));
 					stores.setLocalName(rset.getString("local_name"));
 					stores.setCategoryName(rset.getString("category_name"));
-					stores.setScore(rset.getInt("avg"));
+					stores.setScore(rset.getDouble("avg"));
 					stores.setPrice(rset.getString("price"));
 					stores.setHomepage(rset.getString("homepage"));
 					stores.setImgRoute(rset.getString("photo_n1"));
